+++ option-t.d.ts
@@ -32,6 +32,7 @@ declare module 'option-t' {
         expect(msg: string): T;
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
+        mapOr<U>(def: U, fn: (v: T) => U): U;
         and<U>(optb: Option<U>): Option<U>;
         andThen<U>(fn: (v: T) => Option<U>): Option<U>;
         or(optb: Option<T>): Option<T>;
@@ -48,6 +49,7 @@ declare module 'option-t' {
         expect(msg: string): T;
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
+        mapOr<U>(def: U, fn: (v: T) => U): U;
         and<U>(optb: Option<U>): Option<U>;
         andThen<U>(fn: (v: T) => Option<U>): Option<U>;
         or(optb: Option<T>): Option<T>;
@@ -64,6 +66,7 @@ declare module 'option-t' {
         expect(msg: string): T;
         map<U>(fn: (v: T) => U): Option<U>;
         flatMap<U>(fn: (v: T) => Option<U>): Option<U>;
+        mapOr<U>(def: U, fn: (v: T) => U): U;
         and<U>(optb: Option<U>): Option<U>;
         andThen<U>(fn: (v: T) => Option<U>): Option<U>;
         or(optb: Option<T>): Option<T>;